requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
python-dotenv==1.2.2
requests==2.32.5
openai==2.28.0
python-multipart==0.0.22
reportlab==4.4.10
